refactor: move DateRangePreset to Enum namespace and use DateTimeImmutable
